+++ R/formatage.abreviation.R
@@ -36,7 +36,7 @@
 #####################
 
 formatage.abreviation <- function(
-  thematique = c("Complet", "Écosystème", "Espèce", "Chronique", "Biologie", "Physico-chimie", "Morphologie", "Document", "Institution", "Usage"),
+  thematique = c("Complet", "Écosystème", "Espèce", "Général", "Chronique", "Biologie", "Physico-chimie", "Morphologie", "Document", "Institution", "Usage"),
   formatage = c("Aucun","Propre", "Latex"),
   export = F,
   exportformat = c(".xlsx",".pdf", ".tex")